pyproject.toml
[project]
name = "baru-ai"
version = "0.1.0"
description = "Baru-AI — AI video studio (Electron desktop)"
authors = [{name = "yohomin", email = "bauzonkid@yohomin.com"}]
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.0",
    "certifi>=2025.10.5",
    "openai>=2.6.0",
    "google-genai>=0.3.0",
    "edge-tts==7.2.7",
    "comfykit>=0.1.12",
    "ffmpeg-python>=0.2.0",
    "pillow>=10.0.0,<12",
    "playwright>=1.58.0",
    "beautifulsoup4>=4.14.2",
    "aiohttp>=3.10.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["baru_api", "baru_ai"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
